rename intermediate list files to fee list
