+++ src/apps/fee-list/FeeList.mount.ts
@@ -1,0 +1,4 @@
+import addMount from "../../core/addMount";
+import FeeList from "./FeeList.entry";
+
+addMount({ appName: "fee-list", app: FeeList });
