+++ src/apps/fee-list/modal/fee-details-modal.tsx
