+++ src/apps/fee-list/stackable-fees/fee-details-content.tsx
