--- src/apps/intermediate-list/intermediate-list.mount.ts
@@ -1,4 +1,0 @@
-import addMount from "../../core/addMount";
-import IntermedateList from "./intermediate-list.entry";
-
-addMount({ appName: "intermediate-list", app: IntermedateList });
